Updated Sell Caching, Forms, moved sizing objects
