+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "client",
-  "proxy": "http://34.230.87.84:4000/",
+  "proxy": "http://34.228.14.42:4000/",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
