+++ src/actions/signup.js
@@ -5,7 +5,7 @@ import { setCurrentUser } from './login.js'
 
 export function signup(userData) {
   return dispatch => {
-    return axios.post('http://localhost:4000/api/signup', userData) // Tries to add to DB
+    return axios.post('/api/signup', userData) // Tries to add to DB
     .then(res => {
       const token = res.data.token
       localStorage.setItem('jwtToken', token)
@@ -17,7 +17,7 @@ export function signup(userData) {
 
 export function signupFacebook(userData) {
   return dispatch => {
-    return axios.post('http://localhost:4000/api/signup/facebook', userData)
+    return axios.post('/api/signup/facebook', userData)
     .then(res => {
       const token = res.data.token
       localStorage.setItem('jwtToken', token)
@@ -29,7 +29,7 @@ export function signupFacebook(userData) {
 
 export function signupGoogle(userData) {
   return dispatch => {
-    return axios.post('http://localhost:4000/api/signup/google', userData)
+    return axios.post('/api/signup/google', userData)
     .then(res => {
       const token = res.data.token
       localStorage.setItem('jwtToken', token)
