+++ src/components/ItemImage.jsx
@@ -16,7 +16,7 @@ class ImageCard extends Component {
         {this.state.large ? <Polaroid indx={this.props.indx} src={this.props.src} onChangeDesc={this.props.onChangeDesc} /> : null}
         {this.state.large ? <span className="overshadow" onClick={this.expand} /> : null}
         <div className='imgPreview'>
-          <span className="delImg" onClick={() => this.props.onDel(this.props.src)} src={this.props.src}>Delete</span>
+          <span className="delImg" onClick={() => this.props.onDel(this.props.indx)} src={this.props.src}>Delete</span>
           <span className="expandImg" onClick={this.expand}>Open</span>
           <img src={this.props.src.preview} alt='item' className="imgContent" />
         </div>
