+++ src/components/data/Sizing.js
@@ -1,0 +1,130 @@
+
+
+const sizing = {
+  oneSize: [
+    { value: "one size", label: 'One Size' }
+  ],
+  tops: [
+    { value: "xxs", label: 'XX-Small' },
+    { value: "xs", label: 'X-Small' },
+    { value: "s", label: 'Small' },
+    { value: "m", label: 'Medium' },
+    { value: "l", label: 'Large' },
+    { value: "xl", label: 'X-Large' },
+    { value: "xxl", label: 'XX-Large' },
+  ],
+  kicks: [
+    { value: 6, label: '6' },
+    { value: 6.5, label: '6.5' },
+    { value: 7, label: '7' },
+    { value: 7.5, label: '7.5' },
+    { value: 8, label: '8' },
+    { value: 8.5, label: '8.5' },
+    { value: 9, label: '9' },
+    { value: 9.5, label: '9.5' },
+    { value: 10, label: '10' },
+    { value: 10.5, label: '10.5' },
+    { value: 11, label: '11' },
+    { value: 11.5, label: '11.5' },
+    { value: 12, label: '12' },
+    { value: 12.5, label: '12.5' },
+    { value: 13, label: '13' },
+    { value: 13.5, label: '13.5' },
+    { value: 14, label: '14' },
+    { value: 14.5, label: '14.5' },
+    { value: 15, label: '15' },
+  ],
+  bottoms: [
+    { value: 26, label: '26' },
+    { value: 27, label: '27' },
+    { value: 28, label: '28' },
+    { value: 29, label: '29' },
+    { value: 30, label: '30' },
+    { value: 31, label: '31' },
+    { value: 32, label: '32' },
+    { value: 33, label: '33' },
+    { value: 34, label: '34' },
+    { value: 35, label: '35' },
+    { value: 36, label: '36' },
+    { value: 37, label: '37' },
+    { value: 38, label: '38' },
+    { value: 39, label: '39' },
+    { value: 40, label: '40' },
+    { value: 41, label: '41' },
+    { value: 42, label: '42' },
+    { value: 43, label: '43' },
+    { value: 44, label: '44' },
+  ],
+  accessories: [
+    { value: "xxs", label: 'XX-Small' },
+    { value: "xs", label: 'X-Small' },
+    { value: "s", label: 'Small' },
+    { value: "m", label: 'Medium' },
+    { value: "l", label: 'Large' },
+    { value: "xl", label: 'X-Large' },
+    { value: "xxl", label: 'XX-Large' },
+    { value: 6, label: '6' },
+    { value: 6.5, label: '6.5' },
+    { value: 7, label: '7' },
+    { value: 7.5, label: '7.5' },
+    { value: 8, label: '8' },
+    { value: 8.5, label: '8.5' },
+    { value: 9, label: '9' },
+    { value: 9.5, label: '9.5' },
+    { value: 10, label: '10' },
+    { value: 10.5, label: '10.5' },
+    { value: 11, label: '11' },
+    { value: 11.5, label: '11.5' },
+    { value: 12, label: '12' },
+    { value: 12.5, label: '12.5' },
+    { value: 13, label: '13' },
+    { value: 13.5, label: '13.5' },
+    { value: 14, label: '14' },
+    { value: 14.5, label: '14.5' },
+    { value: 15, label: '15' },
+    { value: 26, label: '26' },
+    { value: 27, label: '27' },
+    { value: 28, label: '28' },
+    { value: 29, label: '29' },
+    { value: 30, label: '30' },
+    { value: 31, label: '31' },
+    { value: 32, label: '32' },
+    { value: 33, label: '33' },
+    { value: 34, label: '34' },
+    { value: 35, label: '35' },
+    { value: 36, label: '36' },
+    { value: 37, label: '37' },
+    { value: 38, label: '38' },
+    { value: 39, label: '39' },
+    { value: 40, label: '40' },
+    { value: 41, label: '41' },
+    { value: 42, label: '42' },
+    { value: 43, label: '43' },
+    { value: 44, label: '44' },
+  ],
+  headwear: [
+    { value: "xxs", label: 'XX-Small' },
+    { value: "xs", label: 'X-Small' },
+    { value: "s", label: 'Small' },
+    { value: "m", label: 'Medium' },
+    { value: "l", label: 'Large' },
+    { value: "xl", label: 'X-Large' },
+    { value: "xxl", label: 'XX-Large' },
+  ]
+}
+
+const categoryOptions = [
+  { value: "tops", label: 'Tops' },
+  { value: "bottoms", label: 'Bottoms' },
+  { value: "kicks", label: 'Kicks' },
+  { value: "accessories", label: 'Accessories' },
+  { value: "headwear", label: 'Headwear' },
+]
+
+const conditionOptions = [
+  { value: "new", label: "New" },
+  { value: "worn", label: "Worn" },
+  { value: "trash", label: "Trash" }
+]
+
+module.exports = { categoryOptions, conditionOptions, sizing }
